issue-472: Added @objc to all enums to allow Obj-C to see the enum types and associated methods.

+++ Sources/iOS/NavigationDrawerController.swift
@@ -31,7 +31,7 @@
 import UIKit
 
 @objc
-public enum NavigationDrawerPosition: NSInteger {
+public enum NavigationDrawerPosition: Int {
 	case left
 	case right
 }
